refactor(providers): relocate createBlockedAuthorFilter, make app_providers a pure barrel
Move createBlockedAuthorFilter into moderation_providers.dart — its only
dependencies (contentBlocklistRepositoryProvider, contentPolicyEngineProvider)
already live there — and reduce app_providers.dart to export-only. This
removes the last symbol that forced split leaf modules to import the
aggregator back, so the aggregator no longer participates in a dependency
cycle.

moderation_providers itself now imports auth/repository directly instead of
the aggregator (dropping the feedAspectRatio hide that only existed to avoid
the barrel re-export collision).

Refs #4554.
